# src/main/api/api.yaml
openapi: 3.0.1
info:
  title: API
  description: API description in Markdown.
  version: 1.0.0
servers:
  - url: 'localhost:8080'
paths:
...
  /decks:
    post:
      tags:
        - Decks
      summary: "Create deck"
      description: "Language and name is required when sharedDeckId is not provided"
      operationId: createDeck
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeckCreateRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deck"
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Not found
    get:
      tags:
        - Decks
      summary: "List decks"
      operationId: listDecks
      parameters:
        - $ref: '#/components/parameters/limitQueryParam'
        - $ref: '#/components/parameters/pageOffsetQueryParam'
        - $ref: '#/components/parameters/isoLanguageCodeQueryParam'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecksPage"
        '400':
          description: Bad request
        '403':
          description: Forbidden
  /decks/{deckId}:
    parameters:
      - $ref: '#/components/parameters/deckIdPathParam'
    get:
      tags:
        - Decks
      summary: "Get deck"
      operationId: getDeck
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deck"
        '403':
          description: Forbidden
        '404':
          description: Not found
    put:
      tags:
        - Decks
      summary: "Update deck"
      operationId: updateDeck
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeckUpdateRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deck"
        '400':
          description: Bad request
        '403':
          description: Forbidden
    delete:
      tags:
        - Decks
      summary: "Delete deck"
      operationId: deleteDeck
      responses:
        "204":
          description: No content
        '403':
          description: Forbidden
        '404':
          description: Not found
  /shared/decks:
    post:
      tags:
        - Shared Decks
      summary: "Create shared deck"
      operationId: createSharedDeck
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SharedDeckCreateRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedDeck"
        '400':
          description: Bad request
        '403':
          description: Forbidden
    get:
      tags:
        - Shared Decks
      summary: "List shared decks"
      operationId: listSharedDecks
      parameters:
        - $ref: '#/components/parameters/limitQueryParam'
        - $ref: '#/components/parameters/pageOffsetQueryParam'
        - $ref: '#/components/parameters/isoLanguageCodeQueryParam'
        - $ref: '#/components/parameters/sharedDecksSortByQueryParam'
        - $ref: '#/components/parameters/sharedDecksTagFilterQueryParam'
        - $ref: '#/components/parameters/sharedDecksNameFilterQueryParam'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedDecksPage"
        '400':
          description: Bad request
        '403':
          description: Forbidden
  /shared/decks/{sharedDeckId}:
    parameters:
      - $ref: '#/components/parameters/sharedDeckIdPathParam'
    get:
      tags:
        - Shared Decks
      summary: "Get shared deck"
      operationId: getSharedDeck
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedDeck"
        '403':
          description: Forbidden
        '404':
          description: Not found
    put:
      tags:
        - Shared Decks
      summary: "Update shared deck"
      operationId: updateSharedDeck
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SharedDeckUpdateRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedDeck"
        '400':
          description: Bad request
        '403':
          description: Forbidden
    delete:
      tags:
        - Shared Decks
      summary: "Delete shared deck"
      operationId: deleteSharedDeck
      responses:
        "204":
          description: No content
        '403':
          description: Forbidden
        '404':
          description: Not found
  /shared/decks/{deckId}/cards:
    parameters:
      - $ref: '#/components/parameters/deckIdPathParam'
    get:
      tags:
        - Shared Decks
      summary: "List shared cards"
      operationId: listSharedCards
      parameters:
        - $ref: '#/components/parameters/cardsLimitQueryParam'
        - $ref: '#/components/parameters/pageOffsetQueryParam'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedCardsPage"
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Not found
  /decks/{deckId}/statistics:
    parameters:
      - $ref: '#/components/parameters/deckIdPathParam'
    get:
      tags:
        - Decks
      summary: "Get deck statistics"
      operationId: getDeckStatistics
      parameters:
        - $ref: '#/components/parameters/statisticsIntervalQueryParam'
        - $ref: '#/components/parameters/statisticsDateRangeStartQueryParam'
        - $ref: '#/components/parameters/statisticsDateRangeEndQueryParam'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeckStatistics"
        '403':
          description: Forbidden
        '404':
          description: Not found
  /decks/{deckId}/cards:
    parameters:
      - $ref: '#/components/parameters/deckIdPathParam'
    post:
      tags:
        - Cards
      summary: "Create card"
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardCreateRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        '400':
          description: Bad request
        '403':
          description: Forbidden
    get:
      tags:
        - Cards
      summary: "List cards"
      operationId: listCards
      parameters:
        - $ref: '#/components/parameters/cardsLimitQueryParam'
        - $ref: '#/components/parameters/pageOffsetQueryParam'
        - $ref: '#/components/parameters/cardSortByQueryParam'
        - $ref: '#/components/parameters/cardShowOnlyQueryParam'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardsPage"
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Not found
  /decks/{deckId}/cards/{cardId}/review-answers:
    parameters:
      - $ref: '#/components/parameters/deckIdPathParam'
      - $ref: '#/components/parameters/cardIdPathParam'
    post:
      tags:
        - Cards
      summary: "Submit answer"
      operationId: submitAnswerForCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmittedCardReviewAnswer'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledCardReviews"
        '400':
          description: Bad request
        '403':
          description: Forbidden
        '404':
          description: Not found
  /decks/{deckId}/cards/{cardId}:
    parameters:
      - $ref: '#/components/parameters/deckIdPathParam'
      - $ref: '#/components/parameters/cardIdPathParam'
    patch:
      tags:
        - Cards
      summary: "Update card"
      operationId: patchCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardUpdateRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        '400':
          description: Bad request
        '403':
          description: Forbidden
    get:
      tags:
        - Cards
      summary: "Get card"
      operationId: getCard
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        '403':
          description: Forbidden
        '404':
          description: Not found
    delete:
      tags:
        - Cards
      summary: "Delete card"
      operationId: deleteCard
      responses:
        "204":
          description: No content
        '403':
          description: Forbidden
        '404':
          description: Not found
...
components:
  schemas:
...
    PageBase:
      type: object
      properties:
        limit:
          $ref: '#/components/schemas/Limit'
        currentPage:
          $ref: '#/components/schemas/PageOffset'
        pageTotal:
          $ref: '#/components/schemas/PageTotal'
        total:
          $ref: '#/components/schemas/ItemsTotal'
    Limit:
      type: integer
      format: int32
      description: "Current limit of items per page"
    PageOffset:
      type: integer
      format: int32
      minimum: 0
      default: 0
      description: "Zero-based page number"
    PageTotal:
      type: integer
      format: int32
      description: "Total number of pages"
    ItemsTotal:
      type: integer
      format: int32
      description: "Total number of items"
...
    UserDeckSettings:
      type: object
      properties:
        newCardsPerDay:
          $ref: '#/components/schemas/NewCardsPerDaySetting'
        easyRate:
          $ref: '#/components/schemas/EasyRateSetting'
        hardRate:
          $ref: '#/components/schemas/HardRateSetting'
        maxInterval:
          $ref: '#/components/schemas/MaxIntervalSetting'
        minInterval:
          $ref: '#/components/schemas/MinIntervalSetting'
        globalEaseModifier:
          $ref: '#/components/schemas/GlobalEaseModifierSetting'
        intervalRateAfterFail:
          $ref: '#/components/schemas/IntervalRateAfterFailSetting'
        leechThreshold:
          $ref: '#/components/schemas/LeechThresholdSetting'
        newCardSteps:
          $ref: '#/components/schemas/NewCardStepsSetting'
...
    MaxIntervalSetting:
      type: integer
      format: int32
      minimum: 1
      description: "The maximum interval (in days) that can be used in determining the date of the next review"
    IntervalRateAfterFailSetting:
      type: number
      format: float
      minimum: 0
      exclusiveMinimum: true
      description: "Multiplier used in calculating the date of the next review if the threshold of incorrect answers is exceeded"
...
    NewCardsPerDaySetting:
      type: integer
      format: int32
      minimum: 0
      description: "The number of new cards that should appear during review"
    EasyRateSetting:
      type: number
      format: float
      minimum: 1
      description: "The multiplier for the 'easy' answer, used during calculation of the next review date"
    HardRateSetting:
      type: number
      format: float
      minimum: 0
      exclusiveMinimum: true
      maximum: 1
      description: "The multiplier for the 'hard' answer, used during calculation of the next review date"
    MinIntervalSetting:
      type: integer
      format: int32
      minimum: 1
      description: "The minimum interval (in days) that can be used in determining the date of the next review"
    GlobalEaseModifierSetting:
      type: number
      format: float
      minimum: 0
      exclusiveMinimum: true
      description: "The multiplier used during calculation of the next review date for correct answers"
    LeechThresholdSetting:
      type: integer
      format: int32
      minimum: 1
      description: "Threshold of failed answers after which card is marked as leech"
    NewCardStepsSetting:
      type: array
      minItems: 1
      description: "For each new card, you must give as many correct answers in a row as the number of steps defined"
      items:
        $ref: '#/components/schemas/NewCardStepSetting'
...
    Deck:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/DeckId'
        name:
          $ref: '#/components/schemas/DeckName'
        sharedDeckId:
          $ref: '#/components/schemas/OriginSharedDeckId'
        language:
          $ref: '#/components/schemas/IsoLanguageCode'
        customSettings:
          $ref: '#/components/schemas/UserDeckSettings'
        effectiveSettings:
          $ref: '#/components/schemas/UserDeckSettings'
    DeckCreateRequest:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/DeckName'
        sharedDeckId:
          $ref: '#/components/schemas/OriginSharedDeckId'
        language:
          $ref: '#/components/schemas/IsoLanguageCode'
    DeckUpdateRequest:
      type: object
      required:
        - name
        - language
      properties:
        name:
          $ref: '#/components/schemas/DeckName'
        language:
          $ref: '#/components/schemas/IsoLanguageCode'
        customSettings:
          $ref: '#/components/schemas/UserDeckSettings'
    DecksPage:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          properties:
            decks:
              type: array
              items:
                $ref: '#/components/schemas/DeckSearchResult'
    DeckId:
      type: string
    DeckSearchResult:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/DeckId'
        name:
          $ref: '#/components/schemas/DeckName'
        sharedDeckId:
          $ref: '#/components/schemas/OriginSharedDeckId'
        language:
          $ref: '#/components/schemas/IsoLanguageCode'
        waitingReviews:
          $ref: '#/components/schemas/DeckWaitingReviews'
    DeckName:
      type: string
      maxLength: 100
    SharedDeckId:
      type: string
    OriginSharedDeckId:
      type: string
      description: "Id of the shared deck that was used to create this deck"
    IsoLanguageCode:
      type: string
      description: "ISO 639-1 code"
    CardId:
      type: string
    CardCreateRequest:
      type: object
      required:
        - front
        - back
      properties:
        front:
          $ref: '#/components/schemas/CardContentElements'
        back:
          $ref: '#/components/schemas/CardContentElements'
    Card:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/CardId'
        front:
          $ref: '#/components/schemas/CardContentElements'
        back:
          $ref: '#/components/schemas/CardContentElements'
        status:
          $ref: '#/components/schemas/CardStatus'
        nextReview:
          $ref: '#/components/schemas/CardNextReview'
        easeRate:
          $ref: '#/components/schemas/CardEaseRate'
        interval:
          $ref: '#/components/schemas/CardInterval'
        suspended:
          $ref: '#/components/schemas/CardSuspended'
        currentStep:
          $ref: '#/components/schemas/CardCurrentStep'
        statistics:
          $ref: '#/components/schemas/CardStatistics'
        leech:
          $ref: '#/components/schemas/CardLeech'
    CardsPage:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          properties:
            cards:
              type: array
              items:
                $ref: '#/components/schemas/Card'
    CardUpdateRequest:
      type: object
      properties:
        front:
          $ref: '#/components/schemas/CardContentElements'
        back:
          $ref: '#/components/schemas/CardContentElements'
        suspended:
          $ref: '#/components/schemas/CardSuspended'
        leech:
          $ref: '#/components/schemas/CardLeech'
    CardContentElements:
      type: array
      minItems: 1
      items:
        $ref: '#/components/schemas/CardContentElement'
    CardContentElement:
      type: object
      required:
        - type
        - content
      properties:
        type:
          $ref: '#/components/schemas/CardContentElementType'
        content:
          $ref: '#/components/schemas/CardContentElementContent'
...
    CardStatus:
      type: string
      description: "Cards go from REVIEWED_YOUNG to REVIEWED_OLD after the interval is bigger than 30 days"
      enum:
        - NOT_SEEN
        - REVIEWED_YOUNG
        - REVIEWED_OLD
    CardNextReview:
      type: string
    CardEaseRate:
      type: number
      format: float
    CardInterval:
      type: integer
      format: int32
      description: "Current interval in days"
    CardSuspended:
      type: boolean
    CardCurrentStep:
      type: integer
      format: int32
      description: "Current step counting from 0"
    CardStatistics:
      type: object
      properties:
        fails:
          $ref: '#/components/schemas/CardFailsStatistic'
        reviews:
          $ref: '#/components/schemas/CardReviewsStatistic'
    CardLeech:
      type: boolean
      description: "Indicates whether the card is a leech (Leech is a card which should be improved as user was not able to memorize after x attempts)"
    CardFailsStatistic:
      type: integer
      format: int32
...
    DateRangeParam:
      type: string
      description: "Date in YYYY-MM-DD format, if interval interval is set to MONTHLY then days are ignored"
      example: "2023-09-22"
    DeckStatistics:
      type: object
      properties:
        statistics:
          $ref: '#/components/schemas/DeckStatisticsEntries'
    DeckStatisticsEntries:
      type: array
      items:
        $ref: '#/components/schemas/DeckStatisticsEntry'
    DeckStatisticsEntry:
      type: object
      properties:
        index:
          $ref: '#/components/schemas/DeckStatisticsEntryIndex'
        hard:
          $ref: '#/components/schemas/DeckStatisticsEntryHard'
        normal:
          $ref: '#/components/schemas/DeckStatisticsEntryNormal'
        easy:
          $ref: '#/components/schemas/DeckStatisticsEntryEasy'
        wrong:
          $ref: '#/components/schemas/DeckStatisticsEntryWrong'
        suspended:
          $ref: '#/components/schemas/DeckStatisticsEntrySuspended'
...
    ScheduledCardReviews:
      type: object
      properties:
        total:
          $ref: '#/components/schemas/ItemsTotal'
        cards:
          type: array
          items:
            $ref: '#/components/schemas/CardId'
        cardToReview:
          $ref: '#/components/schemas/Card'
    DeckWaitingReviews:
      type: integer
      format: int32
      description: "Number of reviews scheduled for today"
    SharedDeckCreateRequest:
      type: object
      required:
        - name
        - language
      properties:
        deckId:
          $ref: '#/components/schemas/DeckId'
        name:
          $ref: '#/components/schemas/DeckName'
        language:
          $ref: '#/components/schemas/IsoLanguageCode'
        tags:
          $ref: '#/components/schemas/SharedDeckTags'
        description:
          $ref: '#/components/schemas/SharedDeckDescription'
    SharedDeck:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/SharedDeckId'
        ownerUsername:
          $ref: '#/components/schemas/SharedDeckOwnerUsername'
        sharedAt:
          $ref: '#/components/schemas/SharedDeckSharedAtDate'
        name:
          $ref: '#/components/schemas/DeckName'
        language:
          $ref: '#/components/schemas/IsoLanguageCode'
        tags:
          $ref: '#/components/schemas/SharedDeckTags'
        description:
          $ref: '#/components/schemas/SharedDeckDescription'
        cardCount:
          $ref: '#/components/schemas/SharedDeckCardCount'
        popularity:
          $ref: '#/components/schemas/SharedDeckPopularity'
    SharedDeckOwnerUsername:
      type: string
    SharedDeckSharedAtDate:
      type: string
    SharedDeckTags:
      type: array
      maxItems: 30
      items:
        $ref: '#/components/schemas/SharedDeckTag'
    SharedDeckTag:
      type: string
      maxLength: 100
    SharedDeckDescription:
      type: string
      maxLength: 100000
    SharedDeckCardCount:
      type: integer
      format: int32
    SharedDeckPopularity:
      type: integer
      format: int32
    SharedDecksPage:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          properties:
            decks:
              type: array
              items:
                $ref: '#/components/schemas/SharedDeckSearchResult'
    SharedDeckUpdateRequest:
      type: object
      required:
        - name
        - language
      properties:
        name:
          $ref: '#/components/schemas/DeckName'
        language:
          $ref: '#/components/schemas/IsoLanguageCode'
        tags:
          $ref: '#/components/schemas/SharedDeckTags'
        description:
          $ref: '#/components/schemas/SharedDeckDescription'
    SharedDeckSearchResult:
      type: object
      properties:
        sharedAt:
          $ref: '#/components/schemas/SharedDeckSharedAtDate'
        name:
          $ref: '#/components/schemas/DeckName'
        language:
          $ref: '#/components/schemas/IsoLanguageCode'
        tags:
          $ref: '#/components/schemas/SharedDeckTags'
        cardCount:
          $ref: '#/components/schemas/SharedDeckCardCount'
        popularity:
          $ref: '#/components/schemas/SharedDeckPopularity'
    SharedCard:
      type: object
      properties:
        front:
          $ref: '#/components/schemas/CardContentElements'
        back:
          $ref: '#/components/schemas/CardContentElements'
    SharedCardsPage:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          properties:
            cards:
              type: array
              items:
                $ref: '#/components/schemas/SharedCard'
...
    SubmittedCardReviewAnswer:
      type: object
      properties:
        answer:
          $ref: '#/components/schemas/CardReviewAnswer'
    CardReviewAnswer:
      type: string
      enum:
        - WRONG
        - HARD
        - GOOD
        - EASY
        - SUSPEND
...
  parameters:
...
    isoLanguageCodeQueryParam:
      name: language
      in: query
      description: "ISO 639-1 code to filter by"
      required: false
      schema:
        $ref: '#/components/schemas/IsoLanguageCode'
    pageOffsetQueryParam:
      name: offset
      in: query
      description: "Zero-based page number"
      required: false
      schema:
        $ref: '#/components/schemas/PageOffset'
    limitQueryParam:
      name: limit
      in: query
      description: "The maximum number of entries to return"
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 100
    cardsLimitQueryParam:
      name: limit
      in: query
      description: "The maximum number of entries to return"
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 50
        default: 20
...
    deckIdPathParam:
      name: deckId
      in: path
      description: 'Unique ID of the deck'
      required: true
      schema:
        $ref: '#/components/schemas/DeckId'
    sharedDeckIdPathParam:
      name: deckId
      in: path
      description: 'Unique ID of the shared deck'
      required: true
      schema:
        $ref: '#/components/schemas/SharedDeckId'
    cardIdPathParam:
      name: cardId
      in: path
      description: 'Unique ID of the card'
      required: true
      schema:
        $ref: '#/components/schemas/CardId'
    cardSortByQueryParam:
      name: sortBy
      in: query
      required: false
      schema:
        type: string
        enum:
          - NEXT_REVIEW_ASC
          - NEXT_REVIEW_DESC
          - INTERVAL_ASC
          - INTERVAL_DESC
    cardShowOnlyQueryParam:
      name: showOnly
      in: query
      required: false
      schema:
        type: string
        enum:
          - LEECH
          - SUSPENDED
          - NOT_SEEN
          - REVIEWED_YOUNG
          - REVIEWED_OLD
    statisticsIntervalQueryParam:
      name: interval
      in: query
      required: true
      description: |
        The interval at which statistics should be presented. If date range query param is not provided then API will return statistics for:
        * MONTHLY = last 12 months 
        * DAILY = all days of the current month
      schema:
        type: string
        enum:
          - MONTHLY
          - DAILY
    statisticsDateRangeStartQueryParam:
      name: dateRangeStart
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/DateRangeParam'
    statisticsDateRangeEndQueryParam:
      name: dateRangeEnd
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/DateRangeParam'
    sharedDecksSortByQueryParam:
      name: sortBy
      in: query
      required: false
      schema:
        type: string
        enum:
          - CARD_COUNT_ASC
          - CARD_COUNT_DESC
          - POPULARITY_ASC
          - POPULARITY_DESC
    sharedDecksTagFilterQueryParam:
      name: tag
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/SharedDeckTag'
    sharedDecksNameFilterQueryParam:
      name: name
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/DeckName'